requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
neo4j==5.14.0
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.25.0
flake8==6.1.0
mypy==1.6.1
safety==2.3.5
bandit==1.7.5
